requirements.txt
blue==0.9.1
Django==4.2.4
django-anymail[mailjet]==8.6
django-cors-headers==4.0.0
django-environ==0.10.0
django-filter==23.2
django-storages[google]==1.13.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.26.2
drf-yasg==1.21.7
gunicorn==20.1.0
Pillow==9.3.0
pkce==1.0.3
pre-commit==3.3.3
psycopg==3.1.10
psycopg2-binary==2.9.7
PyJWT==2.3.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-django==4.5.2
python-decouple==3.6
requests==2.31.0
ruff==0.0.282
sentry-sdk==1.39.2